+++ src/common/CacheableScriptFile.h
@@ -8,7 +8,7 @@
 
 #include <string>
 #include <vector>
-#include "sphere_library/CSFile.h"
+#include "sphere_library/CSFileText.h"
 
 
 class CacheableScriptFile : public CSFileText
@@ -28,14 +28,14 @@ private:
 public:
 	virtual bool IsFileOpen() const;
 	virtual bool IsEOF() const;
-	virtual tchar *ReadString(tchar *pBuffer, size_t sizemax);
-	virtual size_t Seek(size_t offset = 0, int origin = SEEK_SET);
-	virtual size_t GetPosition() const;
+	virtual tchar * ReadString(tchar *pBuffer, int sizemax);
+	virtual int Seek(int offset = 0, int origin = SEEK_SET);
+	virtual int GetPosition() const;
 
 private:
 	bool m_closed;
 	bool m_realFile;
-	size_t m_currentLine;
+	int m_currentLine;
 
 protected:
 	std::vector<std::string> m_fileContent;
